[feat/daengle-91] 댕글미터 계산 로직 구현 완료

+++ daengle-user-api/src/main/java/ddog/user/presentation/review/dto/request/UpdateGroomingReviewInfo.java
@@ -9,7 +9,7 @@ import java.util.List;
 @Builder
 @Getter
 public class UpdateGroomingReviewInfo {
-    private Long starRating;
+    private Integer starRating;
     private List<GroomingKeyword> groomingKeywordList;
     private String content;
     private List<String> imageUrlList;
